Starting with tabs
The tab system was sorely lacking last time, with it being tightly integrated, there wasnt much room for other things in the tabs.

The new tabs system takes a lot of the code from the old system, but makes the tab bar its own seperate component and keeps the tabs in their own svelte store allowing for more tab options in the future and tabs are no longer required to have an editor. Since this allows for tabs to have other elements in the future, such as a markdown previewer