Working Keymap system
The keymap system is now working.

The keymap is split up in 2 parts.

The first part can be found in the add-cm-keybinds file, which creates additional commands for the codemirror 6 editor

The second part can be found in the keymap file, which adds **global** UI shortcuts to the editor, the event listener is located in the TabBar component, but this will likely be changed in a future update