Terminal Fixes
Fixed an issue where the terminal rows and columns would be miscalculated by the fit addon, which caused the font to looked stretched.

the fit_terminal function waits until the next event cycle to resize the terminal which solves the font issue

Fixed an issue where the hidden css class applied to the terminal stopped users from seeing the terminals output, the css class now uses transforms display instead of height, which looks worse, but does fix the issue.

This will likely be worked on more since there is a lot more broken with the terminal system as is