Register Language Extensions
Expanded on the language system, so that registering a language also means providing languagesupport to the editor.

Registering a new active or global extension updates the editor.

Languages whose file extension have a language support(currently only python with py as the file extension) will automatically have their language set and their language support added to the editor the file being edited in.

New file extensions for languages can be easily registered with the add_file_association function, but this only works if the language already has a registered language support extension and at least 1 file extension associated with that language