windows fixes
fixed a regression in the decoding of URI's on windows

fixed the path normalization so that paths that only differ by capitalization. This was done to make the paths on windows the same since the drive letter from the uri is stored lowercase while the paths are uppercase. This commit fixes this by just capitalizing the drive letter if it is a windows path