LSP: didChange notification
Added a codemirror extension that send every keystroke to the lsp using the textDocument/didChange notification. It currently does this one key stroke at a time, but later I plan to use a short delay to send multiple keystrokes at once.

There is currently a bug that i just cannot figure out and that is that a publish diagnostics notification is only sent/received from the language-server when a request is sent, which is bad and it means we have to actively send requests for updates